requirements.txt
decorator==4.0.11
Django==1.11.1
django-extensions==1.7.9
django-filter==1.0.2
djangorestframework==3.6.2
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
wheel==0.24.0
